WIP: refactored service clients location
