+++ src/helpers/uploads/upload-to-cloudinary.js
@@ -1,16 +1,10 @@
-const ClodinaryClient = require('cloudinary').v2;
+const client = require('../client/cloudinary-client');
 const _ = require('lodash');
 const Promise = require('bluebird');
 const httpError = require('http-errors');
-const config = require('smart-config').get('cloudinary');
 
 module.exports = async ({ input }) => {
-  ClodinaryClient.config({
-    cloud_name: config.cloud_name,
-    api_key: config.api_key,
-    api_secret: config.api_secret
-  });
-  return uploadMany(ClodinaryClient, input);
+  return uploadMany(client, input);
 };
 
 function upload(client, file) {
